chain: refactor assign_with_possible_repeats in shard_assignment.rs (#6253)
Restructure the code a little such that `shard_index` is popped/peeked
in only one location and the ‘if→loop→if+else-if+else-if+…’ structure
is simplified to ‘loop→match’ which is hopefully slightly easier to
comprehend.

While at it, use `peek` when accessing the `shard_index` heap.  This
way there’s only one place in the code where elements are popped and
added to buffer.  Again, this hopefully simplifies data flow inside of
the function.