neard: make build.rs work with git worktree
The build.rs script for neard package assumes that git direcotry is
located in ../.git (relative to the package directory) which usually
is the case except whet git worktrees are used.  In that case .git is
a file rather than a directory and resolving paths relative to it
acuses an error.  Fix the isuse by using `git rev-parse --show-git-dir`
as the code should from the start.