fix(script): Estimator CI (#6455)
Do not copy runtime_config.json after estimation.

The CI script wants to prepare runtime_configs.json as buildkite
artifact but this is no longer generated by default. Generating it
in an extra step also does not work right now as some estimations are
missing.

We do not need this artifact and this script is deprecated anyway.