fix(chain): don't store to ColTrieChanges in archival nodes
Archival nodes don't need to store to ColTrieChanges because it's only
read when removing unneeded state in clear_block_data(), which
archival nodes never call.

It would be more natural to add an archival param to
ChainStore::new(), but doing things that way would make this a much
bigger change, and much harder to untangle if we ever get rid of
ColTrieChanges. Although it's a bit weird, adding the
ChainStore::set_archival() function allows keeping the change small
and easy to undo.  If/when ColTrieChanges is deleted, we'll want to
basically just revert this change, since then the ChainStore won't
need to know if it's archival anymore.