remove archive_gc_partial_chunks, make it always on
So I’ve been thinking about rolling this feature out and started
wondering about how third party operators will deal with it.  With the
need to run recompress and then setting this option I concluded that
in the end adding the option just leaves window open for someone to
end up with a messy archival node.  On top of that, my current
thinking is to cut 1.25.1 which would act as an opt-in step.