peer_store: apply IP blacklist to peer store (#6503)
When reading peers from the storage mark all peers whose address
is present in the blacklist as banned.  Furthermore, when getting
peers from other peers, ignore all peers that are on our blacklist.

Previously the blacklist was taken into account when peer tried to
establish a new connection with the node.  If a peer was already
stored in the database or another peer sent us the peer, the node
might try to connect to it even if it’s on the blacklist.  With this
change, if a node is on a blacklist we will no longer try to connect
to it.