refactor: Receipt handling to be done in VM logic (#6468)
Refactors receipt handling from being done in all `External` implementations to being done in vm-logic. A few refactoring patterns were explored but this one solved the main issues of testability and deduplication the best I've found.

The tests are somewhat of a property-based test with all common cases included, but this can be expanded to be fuzzed as well. (might do this in separate PR though)