fix!: adjust gas distribution formula for NEP264 (#6536)
Adjusting formula based on @matklad suggestion. It more closely represents distributing gas based on weights if the sum of weights is greater than the amount of gas remaining at the cost of performing slightly more arithmetic.

I think this makes more sense, but hard to think about the exact implications without some benchmark of the two.