perf: force processing of each task as a separate job.
IMPORTANT: light local testing implies that this is not necessary --
tasks are always maximally split. However, I've heard from marcin that
we observe different effect in some real-world experiments, so this
commit is more of an experiment to check if this helps.

To re-iterate, I am 0.7 sure that we shouldn't actually merge this in
master.