refactor: move column definition to a separate file
It takes a lot of code to define all our columns and their properties.
This code gets in a way when reading the rest of the db code.

Moreover, *specific* columns are unrelated to an *abstract* db we
implement in db.rs.

So let's move them into a different file, cleaning up the API a bit in
the process.