feat: separate counter for chunk nodes cache (#6582)
Recent experiments show that cost for reading kv pair from the chunk cache should be somewhere between 37 and 200 Mgas. Given that we use a safety multiplier 3 and that we could build an example where ~1M nodes are read from the chunk cache, we have to add the new cost to the runtime config.

The first step is adding a separate counter for that, and updating trie storage tests appropriately. This PR shouldn't affect the protocol in any way.

Next steps will be:
* add new cost to runtime config with an estimation function
* return to https://github.com/near/nearcore/pull/6450 and charge this cost there

## Testing

* Updating trie storage tests.