feat: rename `contract_compile` gas parameter
The compilation step happens during deployment and is thus charged as
part of `DeployContractAction`. The fee is question is charged when
loading pre-compiled code into the VM for execution. Therefore, it
should be named as such.

This is not a protocol change. It does not change how much gas is
charged nor when it is charged. The only observable difference is the
names used in the gas profiles.