chain: initialise block and header head height metrics at startup
The near_block_height_head and near_header_head_height metrics are
only updated when new block / header is successfully received.  This
means that if a node is syncing headers (and thus not getting any new
blocks) the former metric is not set.  Similarly, if the node takes
a long time to get its first new header or block the corresponding
metric is not set.  This result in the metric not being published.

Initialise the two metrics when Chain is created so that as soon as
node starts up the counters can be monitored.

Fixes: https://github.com/near/nearcore/issues/6409