test: cover gas exceeded cases with function_call_weight
function_call_weight distributes gas at the end of execution. It might
be the case that we run out of either prepaid_gas or max_gas_burnt
limits before that! The two added tests check these two scenarios.

If we run out of `prepaid_gas`, we have nothing to distribute. If we run
out of `max_gas_burnt` though, we still might have some leftover gas to
distribute!

That's OK though: when execution fails for any reason, we don't create
receipts, we do note used_gas, but then we use burnt_gas when creating
refunds.