nearcore: take max_open_files into consideration in recompress_storage
So far recompress_storage assumed that a single database will use 512
file descriptors at the most and thus set NOFILE limit to three times
that.  However, this number is now configurable and may be larger.

Change the command so that it takes into consideration what is the
actual configured max_open_files number.