pytest: fix state_sync_fail.py
this started failing after https://github.com/near/nearcore/pull/6326 because
that PR made it so that the process actually dies when the client actor panics.
Before that change, the node stays up but gives HTTP errors when trying to ask
for stuff that the client actor has to provide. Now that the node will actually exit,
we'll get a ConnectionError instead of an HTTPError