test: use base contract for upgradability tests
The `rs_contract()` uses all the latest protocol features (ie, it
changes over time). This is something you want most of the time, (as we
test against the latest protocol), *but* for upgradability tests
specifically this is wrong -- for those tests, we want to guarantee that
only the base protocol is being used.