integration-tests: fix test_deploy_cost_increased() with set upgrade schedule
this test currently fails when the protocol upgrade schedule for version 53
is set to the future, as it only changes one block's protocol version (at the
beginning of an epoch). The last block produced in the epoch is what will be
taken into account in voting, and it won't take effect for 2 epochs, so fix that
by producing some blocks with the new version advertised for a couple epochs before
proceeding.