metrics: refactor NetworkMetrics and remove bunch of dead functions (#6655)
If creating a counter fails don’t store it in the peer_messages hash
in NetworkMetrics.  There is no point in keeping a None entry in the
hash.  This simplifies inc and inc_by methods which no longer have to
handle the case of None being stored in the hash map (only the case of
value missing from the hash which they were handling already).

With that, inc_counter_opt and inc_counter_by_opt functions are no
longer used.  Just like a whole slew of other functions which
manipulate metrics defined in near_metrics crate.  Delete all of them.