network: remove near_<msg-type>_{total,bytes} metrics
Remove near_<msg-type>_total metrics.  Data in those metrics are
available in a near_peer_message_received_by_type_total{type} metric
which is easier to deal with since it’s a single metric with a label
for all types rather than multiple metrics.

Similarly, replace near_<msg-type>_bytes metrics with a new
near_peer_message_received_by_type_bytes metric.

The near_<msg-type>_dropped metric is left as is for now but will be
dealt with in another change.