Remove usage of strum::AsStaticStr in favour of strum:IntoStaticStr
Convert remaining usage of strum::AsStaticStr (which is deprecated
in never strum versions) with strum:IntoStaticStr which provides
the same feature without the need for a custom trait.

Uses of strum::AsRefStr could also be replaced but they are left as
is for now because conversion is a bit messy.  `msg.as_ref()` would
need to be replaced by `(&msg).into()` and this added borrow is wee
bit ugly.  We may do it in the future commit.