neard: set up rayon thread pool stack size (#6595)
This helps to mitigate the differences in build environments/targes
somewhat. In particular the default thread size is not guaranteed by the
system/libc and it does differ significantly in practice – glibc gives
8MiB by default, whereas musl IIRC allocates 1MiB-large stacks.

I have a feeling that there is a more general pattern to be abstracted away though.
We might want to have some sort of an unified entry-point which handles
printing out the errors, parsing the arguments, setting up the globals (tracing, rayon), etc.