Cached trie node estimation (#6618)
Add an estimation for the new gas cost introduced in #6588.

The cost is called `read_cached_trie_node` and is charged whenever a trie node
is read from the in-memory cache instead of from the database. This is a
numerically small cost but one that is charged often.

The exact cost is hard to measure because it almost exclusively depends on
memory latency, which is different from machine to machine. Plus, the state of
CPU caches drastically influences memory latency as well.

On a high-level, the estimation overcomes the challenges by setting up an
environment that (1) can run an isolated workload that is equivalent to reading
a node from the trie node cache without any overhead, (2) allows to tune
assumptions around memory latency.

Then the estimation measures a sample distribution of the workload for a
pessimistic memory latency assumption.

Using the `--debug` flag of the estimator will also measure the distribution for
other memory assumptions and print a table with the comparison.


Co-authored-by: Longarithm <the.aleksandr.logunov@gmail.com> Co-authored-by:
Aleksandr Logunov <alex.logunov@near.org>