refactor how incoming_receipts are handled in process_block (#6735)
The purpose of this PR is to move the saving of incoming_receipts in process_block to after apply_chunks. In preparation for making apply_chunks non-blocking, process_block shouldn't write any changes to ChainStore before apply_chunks. The PR may seem to make the code more complicated for now, but it will make sense with more following PRs of refactoring process_block.