client: don’t update validators produced counters when syncing
EpochManager::get_validator_info call querying information by block
hash turns out to be quite expensive while node is synchronising.

Normally, EpochManager maintains validator information up to the
latest final block.  With that, calling get_validator_info with
a block hash usually needs to iterate over only a handful of blocks
until it hits the latest final block and then can read the already
aggregated data.

However, while node is syncing, the aggregated data is not maintained.
This means that getting validator info for a given block requires
iteration over all blocks until start of the epoch.  This is an
expensive operation.

EpochManager::get_validator_info is called each time stats line is
generated.  When synchronising this ends up taking a up all of the
processing time to the point that node doesn’t respond to /status
queries.

Stop getting the validator information when node is synchronising to
avoid this problem.  The observable effect is that
- near_validators_blocks_expected,
- near_validators_blocks_produced,
- near_validators_chunks_expected and
- near_validators_chunks_produced
metrics are not updated.