feat: Flat and differential runtime config files (#6693)
Replace the JSON runtime config file per version with a base
`parameters.txt` file plus differential files per version.
The `RuntimeConfig` remains untouched. But it is now loaded from the
flat list of parameters, that converts it to JSON structured data following
the exact layout we use in code, that is then converted to an actual struct
using serde derived code.

The new TXT files are written in a custom format for key-values separated
by a column. Keys must match a defined parameter and values are strings 
or integers. Differential files include old and new values with a `->` between
them.

The direct benefit of this is better readability for config changes and
an easier and less error-prone process of doing such changes.

Indirectly, this could be the first step to further clean up code
around gas costs.
(1) the gas profile could use these parameters to track costs instead
of its own `enum Cost`.
(2) The params-estimator can link estimations to params more easily.

Test plan
---------
The usage of `RuntimeConfigStore` has not changed, only how it is
created. Therefore, tests can ensure all stored `RuntimeConfig`
versions are exactly the same as before. And this should be sufficient
to confirm this change has no side-effects.

To implement this, we keep the old JSON files in the `legacy_configs`
directory. The test `test_old_and_new_runtime_config_format_match` then
iterates through all existing versions and compares the serde
deserialization with configs loaded from TXT files. This is all we need to
verify correctness of the parameters loaded through the new files.

The new TXT file format also comes with custom parsing logic. Therefore,
 a few unit tests to make sure it behaves the way we think in all
circumstances.