use the mock_node framework in a benchmark (#6778)
This sets up sync performance benchmarks using the existing mock_node tool. The basic setup right now is that we check in tarballs containing sample home dirs in the mock_node/benches directory, and then start a node that will sync from a mock network feeding blocks from that home directory, measuring the time it takes for `view_client.send(GetBlock::latest())` to pass a certain height.

Checking in tarballs like this will only be practical for small sample directories, but it should be possible to use the code here to run benchmarks on larger sample directories not checked in to git, possibly with the help of some other tool.

run with `cargo bench -p mock_node -- mock_node_sync_empty` to run the quicker benchmark, or omit the testname to run all.