neard: support no-value --verbose flag
`--verbose` flag is a bit confusing in that it does requires an
argument.  This isn’t contrary to many other Linux utilities which
expect `--verbose` to take no value.

Change the flag such that it accepts an *optional* argument so it can
be passed as plain `--verbose` without any target.

By the way, note that `neard --verbose=<target>` is roughtly equivalent
to `RUST_LOG=info,<target>=debug`.