chain: refactor AdversarialControls to reduce conditional compilation (#6807)
Provide an ‘empty’ definition of AdversarialControls which is used
when test_features compilation feature is not enabled.  This empty
definition provides accessors to all adversarial controls always
returning non-adversarial values.

Based on the test_features flag either the empty definition (if it’s
missing) or definitinons where adversarial behaviour can be enabled
(if it’s present) is used.

With this refactoring most `#[cfg(feature = "test_features")]` guards
can be removed.  This simplifies the code by removing some of the
syntax noise.

While at it, also change the controls such that rather than using
RwLock all the options are stored in atomic values.  This further
simplifies the code slightly.