store: use enum_map::EnumMap in StoreValidator
Similarly to change in RocksDB and TestDB, prefer using EnumMap to
a fixed-size vector in StoreValidator.

Most importantly, this removes dependency on the DBCol::COUNT value
which counts the number of variants ratehr than giving value one more
than the value of largest variant.  With the current code, we could
not remove a deprecated column from the middle of DBCol enum.  Using
EnumMap solves that issue.