test: fix non-determinisit tests
The test might pass or fail depending on whether you `cargo t` or `cargo
t -p integration-tests`.

The reason for that is that `preserve_order` feature of serde_json
differs depending on which cargo command you invoke. That in turn leads
to different input to the contract.

There's two ways we can fix this:

* enableing preserve_order feature for integration tests crate
* formatting the JSON manually

I went with the second option as it seems simpler and more robust.