core: get rid of SharedChunkHeader::prev_block_hash_ref
Change SharedChunkHeader::prev_block_hash method to return a reference
(rather than cloning the hash) so that prev_block_hash_ref method
becomes unnecessary and can be deleted.

Convert all users of all those methods as required.  In a few plasecs
that allowed us to avoid cloning but in others we had to clone either
because we wanted a clone or to satisfy borrow checker.