refactor: avoid platform-specific API in key_file
Permissions::from_mode is portable across unixes and requires u32. But
libc::mode_t isn't guaranteed to be u32, so we have to have this weird
`.into` which might or might not we a no-op, which, eg, confuses clippy.

Instead of relying on platform-specific libc, just specify mode directly
using traditional octals. This allows us to stay fully within std and
drop a dependency from Cargo.toml.