fix: race in NetworkRecepient  (#6828)
closes https://github.com/near/nearcore/issues/6826

We have circular depedencies between PeerManager and Client, which we
break via `dyn PeerManagerAdapter`. As this is a real cycle, we close it
at runtime by calling `set_recipient`. But today nothing prevents client
from sending message to network before that, and indeed that happens if
we are unlucky with timing.

This commit fixes the issue by making `PeerManagerAdapter` APIs block
until the loop is closed.

<hr/>

After finishing the implementation, I realized that this might not be the ideal solution. It seems that actix has the API ([Actor::create](https://docs.rs/actix/latest/actix/prelude/trait.Actor.html#method.create)) for tying such loop nicely, where you can get actor's address before creating the actor. We use it, for example, here: 

https://github.com/near/nearcore/blob/d18f345a7f773131134e3f5ec877b3074fec05ad/integration-tests/src/tests/network/runner.rs#L64-L70

So, in theory, we can just `impl PeerManagerAdapter for Recipient<PeerManagerMessageRequest>` and then rewrite all call-sites to use `create` and friends. 

Upon closer examination, it seems that that is possible, but not trivial, as create API is not orthogonal enough.

* we use `Actor::start_in_arbiter` and `Actor::mock`, which don't have create-like alternatives. We can open-code them, but that's adding more to the framework
* `::create` itself can't return something else, so that means we'd have to use `.unwrap`s

So I think it's better to maybe do just what this PR is doing and just fix the current code. 

Test Plan
---------

Verify that, after applying the following "unfortunate timing" diff

```diff
diff --git a/tools/mock_node/src/setup.rs b/tools/mock_node/src/setup.rs
index 30940262b..9e7493cad 100644
--- a/tools/mock_node/src/setup.rs
+++ b/tools/mock_node/src/setup.rs
@@ -322,6 +322,7 @@ pub fn setup_mock_node(
                 !archival,
             )
         });
+    std::thread::sleep_ms(300);
     network_adapter.set_recipient(mock_network_actor.clone().recipient());
 
     // for some reason, with "test_features", start_http requires PeerManagerActor,
```

The following test

```
$ cargo test --package mock_node --lib -- setup::test::test_mock_node_basic --exact --nocapture
```

fails on master and passes with this PR applied