refactor: new ChainStore API
So far, we are using the following shape of the API for stores:

```rust
// Actual DB with data
struct Store {

}

// DB in the process of being updated.
// StoreUpdate provides the same API as Store, but also takes prospective (not yet committed) changes into account.
// Crucially, it requires exclusive reference to Store.
struct StoreUpdate<'a> {
  store: &'a mut Store,
  foos: HashMap<CryptoHash, Foo>,
  bars: HashMap<CryptoHash, Bar>,
}

impl Store {
  fn store_update<'a>(&'a mut self) -> StoreUpdate<'a> {}
}

impl<'a> StoreUpdate<'a> {
  fn commit(self) -> Result<()>
}
```

The API is to be used like this:

```rust
let mut store = ...;
let mut update = store.store_update();
mutate(&mut update);
update.commit()?;
```

This PR proposes the following alternative shape, which I am going to argue is superior:

```rust
struct Store {

}

struct StoreUpdate<'a> {
  store: &'a Store, // NB: &, not &mut
  diff: StoreDiff,
}

struct StoreDiff {
    foos: HashMap<CryptoHash, Foo>,
    bars: HashMap<CryptoHash, Bar>,
}

impl Store {
  fn store_update<'a>(&'a self) -> StoreUpdate<'a> {}
  fn commit(&mut self, diff: StoreDiff) -> Result<()>
}

impl<'a> StoreUpdate<'a> {
  fn into_diff(self) -> StoreDiff
}
```

The new API is to be used like this:

```rust
let mut store = ...;
let update = prepare_update(&store); // NB: need only & to create update
let diff = update.into_diff();
store.commit(diff)?; // &mut is needed only for commit
```

That is, the change is that creating a store update now requires only a shared reference.
Exclusive reference is needed only when committing the changes. I think this API is better
for the following reasons:

* closer reflects what actually happens -- we are *not* mutating store when working with update.
* is friendlier to parallelism -- now you can be creating several store updates in parallel.
  No silver bullet though -- it's still up to you to sequence the application of updates and ensure
  they are not conflicting (insert-only/rc columns are helpful)
* is more restrictive: now that StoreUpdate doesn't have commit, we significantly reduce the volume
  of code which *could* be committing to the database. In particular, stuff like
  https://near.zulipchat.com/#narrow/stream/322379-pagoda.2Fcore.2Fconcurrent-client/topic/Is.20ChainStoreUpdate.3A.3Atry_save_latest_known.20correct.3F
  becomes impossible by construction