nearcore: use StoreOpener in database migration functions
Use StoreOpener’s encapsulation of the details regarding path to the
RocksDB database.  To make this work deprecate `get_store_version`
method (which takes path to the database as argument) in favour of
a new `StoreOpener::get_version_if_exists` method which, as name
implies, returns version of the database if it exists.

Issue: https://github.com/near/nearcore/issues/6857