Add BlockProcessingPool (#6791)
This PR adds a struct BlockProcessingPool, which will be shared between chain and the future that waits for do_apply_chunks. It already refactors the chain code to use the struct, but doesn't change the chain behavior. In another PR, I will make the call to do_apply_chunks async.