fix(sync): fix a syncing stall when a shard split is happening soon
The client code will actually only update its sync state if
state_sync.run() returns Completed or Changed. If the sync state is
changed in StateSync::sync_shards_status() but the return value is
(false, false), then this won't happen. Note that forgetting to update
update_sync_status only really leads to problems when need_shard and
this_done are both left unset. But can't hurt to just set it to true
whenever we update the status in the code. Concretely, this bug leads
to a stall when syncing state if a shard split is incoming, with log
lines like the following showing up over and over:

022-05-29T17:16:12.216741Z DEBUG sync:run{sync="StateSync"}: sync: syncing state ..... status: StateDownloadComplete ...
...
...
022-05-29T17:16:12.216741Z DEBUG sync:run{sync="StateSync"}: sync: syncing state ..... status: StateDownloadComplete ...
...
...
022-05-29T17:16:12.216741Z DEBUG sync:run{sync="StateSync"}: sync: syncing state ..... status: StateDownloadComplete ...

This is sadly actually pretty hard to add a test for. A unit test
would be nice, but KeyValueRuntime doesn't update its shard layout, so
we can't actually get will_shard_layout_change_next_epoch() to return true
in the test code without adding that functionality. Also just using a
real NightshadeRuntime isn't possible since near-client doesn't import
nearcore, and it feels wrong to import that just for this test...

And then pytests are actually not easy to implement to get this to
fail either, since changing the shard layout in the genesis config and
setting the initial protocol version to 47 (like in
state_sync_fail.py) will cause the shard split to happen at epoch
height 2.  So to trigger this bug we want state sync to happen from
epoch height 1, but there's only actually a small corner case that
will allow that, where we start state sync when our header head is at
epoch height 2, but within 5 blocks of the epoch start. (this is actually
what happened when I ran into the stuck logs above)