rosetta: fix return message when block not final and improve tests (#6966)
get_block_if_final is supposed to return None if block is not found.
However, the way it was written it would return than information as an
error.  This in turn means that Rosetta RPC handlers would report
unknown block as an internal error rather than not found error.  Fix
that.

While at it, improve handling of errors in Rosetta RPC Python test.

This fixes the test which is currently failing on Nayduck.