jsonrpc: refactor `process_request_internal`
Refactor `JsonRpcHandler::process_request_internal` method which is
one of the longest methods in our code base by:
- extracting common code into a new `process_method_call` function,
- moving adversarial requests handling to a new
  `process_adversarial_request_internal` method and
- creating separate methods for each type of adversarial RPC request.

This removes around 100 lines of code and reduces length of the method
from 325 lines to mere 88.

The `process_method_call` function is generic on the argument and
result of an RPC request.  It not only handles bunch of duplicated
code but thanks to type inference it also reduces amount of times the
same type needs to be specified.  The function is actually quite tiny,
but thanks to all the inference and syntax reduction it handles
replaces a few lines each time.

Similarly, introduce `serialise_response` which wraps a common call to
the serde serialisation.  Again, it’s a tiny function but the code it
replaces is a bit noisy so hopefully having a single method makes the
code more readable.

Lastly, `parse_params` was just a duplicate of function of the same
name from `api` module so just use that rather than duplicating the
code.  And observe that `jsonify` has always been called with
`Ok(Ok(value))` argument which means that it can be simplified to the
point where it’s not necessary.