Trace code execution across actix queues for PeerManagerActor->PeerActor messages (#7005)
When an actix actor handles a message from its queue, it does that outside of any spans.
This PR attaches the context of the requester to the request. Thus making the request handling seem like a direct child of spans of the requester.
In the future we'd like to trace execution across process boundaries. For example, to see a partial encoded chunk request made by process A, its handling by process B, and receiving a response by process A again. However, to make this possible we first need to trace execution within the process across the actix system.