fix: make TestDB iterator deterministic (#7029)
Fixes: #7024

Test plan
---------
I did not find a test that could be updated/enhanced to validate that the iteration would remain deterministic.
Not sure whether this extra step is needed. But if it is, I'd appreciate some guidance.

Edit:

Added two tests: `rocksdb_iter_order` and `testdb_iter_order`. These two tests that there is determinism
when calling the `iter` functions.