refactor: drop unused contract preloading infrastructure
The preload is the idea that we usually execute a bunch of contracts in
a row, and that we can pipeline the execution a bit:

while the fist contract is execution, we can, on a separate thread, be
preparing context to run the second contract.

We have this implemented in a proof-of-concept state for about a year,
but this has been dead code still. There's some kernel to the idea, but
it doesn't feel like the most impactful thing we can be working on, and,
realistically, this is going to be left untoched for a long time still.

Specific problems:

* preload doesn't allow us to trivially reduce the cost
* impact from "white-magic" optimizations like 0-copy deser and mmap
  re-use is likely going to be bigger
* we already run some things concurrently, so it's not like other CPUs
  are free. If all the cores are busy, preloading is pure overhead
* we don't *trivially* know the set of contracts we are going to execute
  in a single call to `apply`. It's possible to make efficient use of
  preloading, but that requires complicated code outside of
  near-vm-runner and, potentially, even runtime.