refactor: drop standalone vm runner
The runner tool existed to run a contract as

$ ./standalone-runner code.wasm

It had two potential audiences:

* NEAR developers who are testing the contract
* nearcore developers, who would like to investigate how runtime iteslf
  work.

We've long since deprecated the first usage. Today, `sandbox` is our
only public interface to contract authors, and any such functionality
should be exposed via sandbox.

As for nearcore developers, I've noticed that it's almost always more
convenient to just write a test than to futz with building this
additional binary and its cli arguments.