chain: avoid unnecessary string formatting and copy when reading store
Most existing usage of option_to_not_found results in unnecessary
string formatting or unnecessary copy.  For example in:

    option_to_not_found(
        read_with_cache(&self.store, DBCol::Block, &self.blocks, h.as_ref()),
        &format!("BLOCK: {}", h),
    )

the ‘BLOCK: {}’ string will always be formatted.  If the key exists in
the database that operation will be unnecessary.  On the other hand,
if the key does not exist in the database, option_to_not_found will
copy the value (see `Ok(None)` case):

    match res {
        Ok(Some(o)) => Ok(o),
        Ok(None) => Err(Error::DBNotFoundErr(field_name.to_owned())),
        Err(e) => Err(e.into()),
    }

Avoid this by changing option_to_not_found to take std::fmt::Arguments
argument rather than &str.  This way formatting will happen inside of
the function and only when necessary.