store: various small near_store::db::refcount changes
Firstly, change get_with_rc_logic such that it no longer creates
a new Vec.  Instead, Vec which is passed to it is truncated when
necessary.  This reduces amount of times allocator is called.

Secondly, change decode_value_with_rc so that it returns (None, 0) if
reference count is zero.  This is what the documentation claims it
should do but in reality the function returned (Some(...), 0).

Thirdly, change refcount_merge so that it first goes through all the
values and only at the end performs single Vec allocation.  This makes
merge_refcounted_records function unnecessary and simplifies code
a bunch.

Lastly, rework add_positive_refcount to use [T]::concat which makes
the function much more concise.  (It’s what refcount_merge is already
using).