store: introduce StoreUpdateStorage to StoreUpdate
If tries are set on StoreUpdate, they must point to the same database
as StoreUpdate itself.  This is checked during when commit method is
called.

Change the code such that this condition is encoded in the type.  If
shard tries are set, StoreUpdate does not keep separate reference to
the database.  This means that the check in commit method is no longer
necessary.  The verification is now performed in set_shard_tries and
merge methods.

Furthermore, with this change extend check in merge method to compare
databases in all cases.  Previously, merge would only check that the
ShardTries object match but would otherwise ignore situation where two
StoreUpdate objects referencing different databases would get merged.