network: replace store::Store::new with From implementations (#7131)
Making multiple From implementations makes it easier to create the
object.  Also derive Clone for the type which ends up saving us some
syntax.

The motivation here isn’t obvious and is that in the future it’ll be
easier to add new From implementations (so store::Store can be created
from other arguments) without having to change the call sites.