store: pass StoreUpdate as argument to ShardTries::apply_all
Rather than having ShardTries::apply_all create a new StoreUpdate and
return it, change it such that it takes the update as an argument.
This brings the method in line with other methods of the struct which
also take reference to StoreUpdate as argument.  It also renders some
of the calls to StoreUpdate::merge unnecessary.  The cost is that the
call sites is now a bit more verbose since an explicit creation of the
StoreUpdate is necessary.