Disable the stale bot
Currently, this bot only marks issues as stable but does not actually
close them.  We are not sure if we are prepared to have this bot to
close issues yet.  In general, we are not sure if and how to use this
bot.  So this PR disables the bot entirely.