fix/refactor(mocknet): don't alter forked genesis shard config by default (#7160)
Currently, mocknet.create_genesis_file() alters the genesis file's
shard config in all cases. If the `sharding` argument is true, then it
sets it to the current mainnet config value, and otherwise it sets it
to a single-sharded setup. By default, though, we should not touch it
unless the user asks for that. This has not mattered, really, but when
writing a pytest that forks a localnet this actually does matter,
since we don't want the mainnet config values, and a single shard
isn't what we want, either. We just want the same setup.

This is not quite a no-op. Right now the only user of this function is
in `genesis_updater.py`, which is meant to run on different machines
in a private test network. Before this change, the forked genesis file
will have the same value in both `genesis.shard_layout` and
`genesis.simple_nightshade_shard_layout`, and after, these two will be
different, as in the current mainnet config. But this is probably what
we want, anyway.

Also fix incorrect usage of bool() in genesis_updater.py, since that will
give true for all args except for the empty string